Add encrypted_lvm test
